dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "pandas>=2.3.3",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
